Allow usage of the `generate` action in generators when testing
It was previously not possible to test whether a generator could
invoke another generator using the `generate` command, because this
shells out and can potentially be mocked or its errors can be hidden
from the user. The command always runs, but its effects cannot be
observed. For example, if a file is supposed to be generated using
`generate :model`, the file will not exist in the `destination_root` of
the test when executed. To remedy this, instead of shelling out with
`#run_ruby_script`, use `Generator.invoke` directly and pass in all of
the arguments given to the command.

Fixes #33558