Add or to WhereChain to support simple, flexible conditions
This adds a new method `or` to WhereChain. This can be used to more
easily write WHERE clauses that use the OR operator, using familiar
hash/array/string arguments that `where` takes. But unlike the existing
ActiveRecord::Relation `or`, this will not require repeating the scope.

This is geared towards simpler conditions, where previously the
best option has been passing raw SQL to `where`. If the two (or more)
OR operands are more complex (and maybe have ANDs and ORs inside them)
then the recommendation should still be to define each one as a scope
and use ActiveRecord::Relation `or` to combine them.