Add if_exists option for PG remove_column
It's rather convenient to use if_exists option for drop_table
method in migrations. Unfortunately, this option is not allowed
for remove_column, but it's possible to implement it for PostgreSQL
adapter. SQlite3 allows removing column twice without complaints,
MySQL would still raise an error on second attempt to delete the same column
as it doesn't support this option.