Remove `as_json` from value and tests
Removing `as_json` here revealed that these tests failed because they
return a non-json compliant value of NaNNumber and InfinitNumber.

While we've supported this for awhile one of the issues with this
support is that once we hit `jsonify` all the values _should_ already be
json-compliant values. These tests I removed were added in #26933 - but
in the comments on that PR they were supposed to be deprecated and
removed in the next cycle - but that work never happened. See
https://github.com/rails/rails/pull/26933#issuecomment-260185056

We want to remove this `as_json` here because Rails JSON is very slow.
One of the reasons it's slow is because in some cases we actually loop
over the JSON we're parsing 3 times!

Here are the passes:

1. Call as_json on a tree of objects
  1a. Special objects pass the options down. These are
  Object, Struct, Enumerable, Array, Hash
  1b. This first pass eliminates Structs, Enumerables, and Objects
  Structs replaced with hashes, Enumerables -> Arrays, Objects -> Hash
  1c. New tree contains only: String, Number, Hash, Array, nil, true,
  false
2. Tree is walked again
  2a. Convert Strings to EscapedStrings
  2b. Can retun bad values from as_json (like the infinity and nan)
3. Convert to json

By removing this `as_json` we reduce one of the passes for this case.
The argument here is that the first pass should never have returned a
bad value so we shouldn't have to walk those values a second time - the
value should be `nil` (which it what it will be on pass 3).

[Eileen M. Uchitelle & Aaron Patterson]