Generate expand_cache_key via ActiveSupport::Cache::Key
The current format of the cache key generated via `retrieve_cache_key` which is used by `expand_cache_key` is different than the other method of "expanding" a cache key provided in `expanded_key`. This method is expected to provide a raw cache key that changes as the version changes.

While this method currently relies on the specialized method `cache_key_with_version` this method is not documented in any of our guides and it is confusing that using a custom object with both a `cache_key` and a `cache_version` will not generate an appropriate cache key. 

This method `cache_key_with_version` is also used by the dalli store which resulted in some confusion https://github.com/petergoldstein/dalli/issues/710.

This commit moves all Rails internal cache key generation into the ActiveSupport::Cache::Key object so that if the same objects are used as a key and passed to `expand_cache_key` then it will match an identical key and object being pulled via `expanded_key` if the cache store is `:dalli_store`.

The downsides to this approach:

- The method `cache_key_with_version` now does nothing. Implementing it on an AR object or another custom object will not do anything.
- Anyone relying on this method before will now get cache keys invalidated as the format has changed (instead of `key1-value1/key2-value2` it will now be `key1/key2-value1/value2`).

The upside to this approach:

- Key generation is now completely centralized
- People writing custom objects do not need to know about a third method to implement in addition to `cache_key` and `cache_version`.