Remove tmp directories from yarn workspaces configuration
Having these directories configured as yarn workspaces can cause you to
generate an incorrect root yarn.lock when you run `yarn install` if you
have stale files in these directories. This can be particularly
troubling because the directories are ignored by git so it's not
immediately obvious that you have files there which could be causing an
invalid lockfile to be generated.

Additionally, @lsylvester said that his initial motiviation for
including these directories was resolved by relaxing the requirements
in the package.json template used by the app generator. [1]

[1]: https://github.com/rails/rails/pull/34588#issuecomment-443547937