ActionCable: Same origin match against X-FORWARDED-HOST when proxying
When using `allow_same_origin_as_host` with a reverse proxy the `HOST`
header is set to the proxy host which will cause the CSRF protection
to forbid the request. Each time a request hits a reverse proxy, it
updated the `HOST` header and also prepends the host from the previous
request to the `X-FORWARDED-FOR` header. When the request finally hits
the server where ActionCable is running, the first host in the header
will be the right one for us.

I am proposing to change my original idea introduced in 268c340 to
test against the first element of this header when it is set.