Add `add_global_tags` to ActiveSupport::TaggedLogging.
This is to allow for multiple loggers to be created in a multi-tier architecture.
For example, if one tier handles messaging, another serialization and another
business logic, this allows each tier to have separate loggers that retain
the same tag across all threads. Currently tags are thread-safe and
will be blown away whenever a new thread is created.