Avoid confusion by using string instead of boolean in specs
This commit changes the boolean values used in the specs for strings
to help users to understand how options work.

The `:checked`, `:disabled` and `:readonly` options for
`collection_radio_buttons` are meant to list the elements that should be
affected. However when used with a single boolean value like
`disabled: true` it can be understood as if it was meant to disable the
whole `collection_radio_buttons` helper as it happened here #34915.