Fix error message when adapter is not specified
When we added support for multiple databases through a 3-tiered config
and configuration objects this error message got a bit convoluted.

Previously if you had an application with a missing configuation and
multiple databases the error message would look like this:

```
'doesnexist' database is not configured. Available: development,
development, test, test, production, production
(ActiveRecord::AdapterNotSpecified)
```

That's not very descriptive since it duplicates the environments
(because there are multiple databases per environment for this
application).

To fix this I've constructed a bit more readable error message which now
reads like this if you have a multi db app:

```
The `doesntexist` database is not configured for the `production`
environment. (ActiveRecord::AdapterNotSpecified)

Available databases configurations are:

development: primary, primary_readonly
test: primary, primary_readonly
production: primary, primary_readonly
```

And like this if you have a single db app:

```
The `doesntexist` database is not configured for the `production`
environment. (ActiveRecord::AdapterNotSpecified)

Available databases configurations are:

development
test
```

This makes the error message more readable and presents the user all
available options for the database connections.