Fixed a bug where the debug view does not show the error page properly
There are two cases where the debug view does not show the error details
properly:

 * When the cause is mapped to an HTTP status code the last exception is
   unexpectedly uwrapped
 * When the last error is thrown from a view template the debug view is
   not using the `rescues/template_error.html.erb` to generate the view

Both the cases could be fixed by not unwrapping the exception. The only
case where the exception should be unwrapped is when the last error is
an `ActionView::Template::Error` object. In this case the HTTP status
code is determined based on the cause.

There are actually more wrapper exceptions that are intentionally
thrown. However, there is a consistent pattern of setting the original
message and original backtrace to the wrapper exception implemented, so
the debug view will not lose the information about what went wrong
eariler.