Add ActiveRecord::Relation#take_unique!
Sometimes we make a query expecting a single result, and if for whatever
reason it doesn't it's a sign of data inconsistency or code bug.

Other data access frameworks, like the famous Hibernate (for Java), have
a method that raises an error if the query returns more than one record.
Here's an example:
https://docs.jboss.org/hibernate/orm/3.2/api/org/hibernate/Query.html#uniqueResult()