Improve the type map inizitalizer records processing time
Now the code iterates multiple times in two ways without need:

- First it goes through the records object multiple times doing
partitions. In case we have hundreds of thousands of records (like in
the benchmark I've left below) that means that we have to go
through all of them multiple times doing it very slow. That depends on
the distribution of course but in PG if you have a huge number of tables
the majority of the data are going to be `array_in` or composite
elements
- Second once we have processed the records, we iterate over all of them
again to register every record.

Here is the benchmarks where you can see an slightly improvement for
hundreds of records (the normal case) but if you have a huge number of
types you improves a lot the performance

-- Current method benchmark
Records count: 207
  0.000000   0.010000   0.010000 (  0.002049)
  0.000000   0.000000   0.000000 (  0.002214)
  0.000000   0.000000   0.000000 (  0.001979)
  0.010000   0.000000   0.010000 (  0.002516)
  0.000000   0.000000   0.000000 (  0.002142)
Records count: 142191
  2.050000   0.180000   2.230000 (  2.224884)
  2.090000   0.090000   2.180000 (  2.176774)
  2.050000   0.080000   2.130000 (  2.124951)
  2.140000   0.100000   2.240000 (  2.249662)
  2.040000   0.100000   2.140000 (  2.157886)

-- Suggested method benchmark
Records count: 207
  0.010000   0.000000   0.010000 (  0.001821)
  0.010000   0.000000   0.010000 (  0.001999)
  0.000000   0.000000   0.000000 (  0.001983)
  0.000000   0.000000   0.000000 (  0.001906)
  0.000000   0.000000   0.000000 (  0.001829)
Records count: 142191
  1.050000   0.000000   1.050000 (  1.056097)
  1.040000   0.000000   1.040000 (  1.034397)
  1.020000   0.010000   1.030000 (  1.018424)
  1.100000   0.000000   1.100000 (  1.103647)
  1.030000   0.000000   1.030000 (  1.032034)

I've also checked the objects allocation and they remains practically
the same