Turn lookup context in to a stack, push and pop if formats change
This commit keeps a stack of lookup contexts on the ActionView::Base
instance.  If a format is passed to render, we instantiate a new lookup
context and push it on the stack, that way any child calls to "render"
will use the same format information as the parent.  This also isolates
"sibling" calls to render (multiple calls to render in the same
template).

Fixes #35222 #34138