Refactors a loop
I believe the current style does not clearly communicate that we are
ignoring non-existing autoload paths altogether. Your eyes may even be
looking for an else clause that is easy to miss but that does not exist.

With the early `next` and code comment the loop reads better for my
taste.