Adds support for defined variants and generic support for attachment configuration.
This change allows variants to be named and defined as a fixed list of transformations to apply:

    class User < ApplicationRecord
      has_one_attached :avatar do |attachment|
        attachment.has_variant :thumbnail, resize: "50x50"
      end

      has_many_attached :photos do |attachment|
        attachment.has_variant :rotated, rotated: "-90"
      end
    end

Variants can then be generated by name:

    user.avatar.variant(:thumbnail)

    user.photos.first.variant(:rotated)

As a byproduct of this implementation, this change also allows the `has_*_attached` methods to accept a `&block`. The block yields a configurable object (`ActiveStorage::Attached::Configuration`) which can be used to configure this and any future features as well.