Allow disabling attribute discovery in ActiveRecord
Continuing the work done by @kirs

1. Avoided the incompatibility between ignored_columns and
define_attributes_from_schema=false. You can have an ignored column,
with an attribute with the same name that works the same as attributes
not backed by a column
2. Don't SELECT * when not using necessarily all the columns

I investigated not modifying the columns hash, however it seems this is
just an allow_list to ignored_columns`s disallow_list, and should work
the same way. Perhaps ignored_columns should also not modify the
columns_hash