Add cookies_httponly_default config option.
If set to true, marks all cookies as httpOnly by default.
Can be overridden by setting `httponly: false`.

Cookies that are httpOnly cannot be read by scripts, which increases
security. Session cookies are already marked as httpOnly by default.

This option defaults to false but is set to true in 6.0.