Make Zeitwerk use the ActiveSupport inflector so it inherits any special rules like acroynms
Zeitwork uses a default inflector that’s different than ActiveSupport’s to define which modules can be autoloader based on filesystem paths. If these two inflectors disagree on the filename<->constant conversion, then the modules created for autoloading can be named one thing and the actual constants defined in those files can be named something else. Because the ActiveSupport::Inflector can be customized, it’s possible for the two inflectors to disagree. In Rails 5, only the ActiveSupport::Inflector was used to constantize, so any exceptions applied to it in an initializer or whatever would apply to the autoloader as well. If we make Zeitwork use the ActiveSupport inflector, then they should agree again, and backwards compatibility should be maintained.