Add Procfile by default
The Procfile is a semi-standard that is supported on Heroku, CloudFoundry, Foreman, and more. It is used to communicate with a platform what processes exist that need to be run to boot an application. This PR adds a `Procfile` to be generated by Rails along with a `web` process that instructs platforms how to boot the rails server, and the `release` command which instructs platforms what needs to be done after a build is complete to allow the application to be ready to serve traffic.

The driver of this PR was a conversation with @DHH where he identified a list of items that would allow running Rails on a platform even with fewer steps. Several of these were directly turned into a ticket and have been implemented such as #34710. One of them that was identified was to communicate to a platform that migrations need to be done after a deploy. This auto migration logic can be accomplished using the `release` command in a Procfile.