mysql2 adapter: default datetime attribute to precision = 0
Problem:

Under the mysql2 adapter, when Active Record creates a datetime
attribute without declaration, it puts precision = 0.

https://github.com/rails/rails/blob/382a9563a3bc3dd1982f4db1258e8f0e7f5fac8e/activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb#L586

When an attribute is explicitly declared, like
`attribute :created_at, :datetime`, the precision is set to `nil`
instead.

I beleive declaring `attribute :created_at, :dateimte` should
give the same result as not specifying.

Discussion:

This is a breaking change for users that had their declaration
as `attribute :created_at, :datetime`, because now `created_at`
is subject to rounding.
(See `activemodel/lib/active_model/type/date_time.rb`)

However, the alternative for users that want to switch to explicitly
declaring attribute is putting
`attribute :created_at, precision: 0`, which is not necessary in
other adapters such as `postgresql` and `sqlite`.

I think having precision as 0 is a better default, since it is
consistent with the inferred precision.