Educate in the use of destroy bang
Besides helping product development with automatic code generation,
Generators also are used to educate in how an application in Rails could
be built and how public API should be used.

In terms of public API education I think we are already doing a great
work with the actions `create` and `update` by using the methods
`#create` and `#update` without bang and then taking advantage of their
boolean return value to display an appropriate response. However, at
the same time we are encouraging the use of `#destroy` (without a bang)
without doing anything in case the record can't be destroyed, which
would hide potential bugs. This patch changes that by recommending the
use of `#destroy!`, which raises an exception in case the record can't
be destroyed.