Undefine attr methods on self and descendants when declaring new attribute
Merely reloading the schema from cache can result in a situation where
the model has memoized @attribute_methods_generated, but no method has
actually been defined for the new attribute. It's necessary to undefine
attribute methods on this class and all descendants to avoid situations
where a method for the new attribute may be missing.