Adds named_captures to MatchData to emulate Regex
This change adds a `named_captures` method to
`ActionDispatch::Journey::Path::MatchData` in order to emulate a similar
method present on `Regex`'s `MatchData` present in Ruby core.

This method can be useful for introspection of routes without the need
to use `zip` while testing or developing in Rails core.