Rewrite LazilyDefineAttributes so that it removes itself
LazilyDefineAttributes only needs to stay in the model's ancestors long
enough to define accessors for acceptance attributes, after that it is
not needed anymore. To optimize this, we eagerly define attributes as
soon as either +method_missing+ or +respond_to_missing?+ is called, then
immediately remove those methods from the module so that future calls
will incur no performance penalty.

I've also removed the call to +define_attribute_methods+ here, which
is no longer be necessary to avoid an infinite loop.