Expose all helpers in ActionController#helpers
Traditionally when someone needed the url_helpers in the model the suggested way, as in the docs, is to include Rails.application.routes.url_helpers in the model, i.e:

  class User < ActiveRecord::Base
    def base_uri
       user_path(self)
    end
  end

The problem with this approach is that we include the entirety of the url_helpers in our model interface, and end up mixing view and model concerns.

A better approach is to not include the module but directly call the method in Rails.application.routes.url_helpers like:

def base_uri
  Rails.application.routes.url_helpers.user_path(self)
end

The problem with this approach is this is quite a handful to write and this level of chaining requires a lot coupling to the intermediate structures.

What this PR is proposing is to expose all the helpers including the url_helpers through the existing helpers method like:

ApplicationController.helpers.user_path