Fixes ActiveRecord::MismatchedForeignKey error message when `create_table` fails because a foreign key references a table that does not exist.
Previous behavior:
The very first foreign key of the SQL query is extracted with a regex.
The error message says that the type of this foreign key does not match the
targeted primary key column's type.

The problem is that the very first foreign key of the SQL query is not always
the erroneous one. Also, the error might be due to the targeted table not
existing.

We now loop on each foreign key of the `create_table` statement. If one targets
a table that does not exist, we build an appropriate error message mentionning
the name of the missing table. Otherwise we fallback to the previous behavior.