Remove reminiscences from previous documentation [ci skip]
In #35861 documentation for `after_save_commit` was added, but also old
documentation was adapted. I think adapting the old documentation is not
necessary because there are already examples for `after_commit` on
`:destroy` above. On top of that the action invoked by the callback on
`:destroy` (`:log_user_saved_to_db`) talks about saving when the object
is destroyed, which can be misleading.

As I believe the documentation added by #35861 is already enough this
patch removes the modified part, which made sense before adding docs for
`after_save_commit` but is already covered by other parts of this guide.