Bump RuboCop to 0.68.1
Performance cops were extracted in rubocop-hq/rubocop#5977 and @koic
updated to 0.67.2 and added the rubocop-performance gem in #35989.

This bumps rubocop to 0.68.1 to be on the latest version. Two changes
were made in regards to the bump of the version:
* The first was that Layout/FirstParameterIndentation was renamed to
Layout/IndentFirstArgument. These changes can be seen in
rubocop-hq/rubocop#6982 and rubocop-hq/rubocop#6987
* The second was that a new instance of the Style/HashSyntax cop was
found. I updated that instance to follow the new Ruby >= 1.9 syntax.