Use adapter name from existing connection
When loading the schema we call the `ActiveRecord::Base.connection` to
retrieve the `adpater_name` for an attribute. In a multi-database world
this can break as the resulting connection pool may not exist.

This patch uses the existing connection object when fetching the
adapter_name instead of calling out to `ActiveRecord::Base`.

Fixes: 36139