Only build middleware proxy when instrumentating
The instrumentation proxy adds three stack frames per-middleware, even
when nothing is listening.

This commit, when the middleware stack is built, only adds
instrumentation when the `process_middleware.action_dispatch` event has
already been subscribed to.

The advantage to this is that we don't have any extra stack frames in
apps which don't need middleware instrumentation.

The disadvantage is that the subscriptions need to be in place when the
middleware stack is built (during app boot). I think this is likely okay
because temporary AS::Notifications subscriptions are strongly
discouraged.