Merge payload for EventObject subscribers
When instrumenting a block of code like:

```ruby
ActiveSupport::Notifications.instrument("process_action.action_controller", raw_paylaod) do |payload|
 payload[:view_runtime] = render_view
end
```
If we use an evented subscriber like so:

``` ruby
ActiveSupport::Notifications.subscribe("process_action.action_controller", raw_payload) do |event|
 assert event.payload[:view_runtime]
end
```

The code breaks because the underlying EventObject's payload does not have the
`:view_runtime` key added during instrumentation.

This is because the `EventedObject` subscriber calls the `finish` method with the
`payload` of the event at the time it was pushed into the stack, before the
block executes, but we want to call `finish`  with the `payload` after the
instrument block executes this way if the `payload` was modified during the block
we have access to it. This is consistent with the other types of subscribers
who don't have this bug.