Prevent double save of has_one association when saving child
Autosave double saves a has_one association when saving the child first.
The second save would clear the mutation tracker resulting in an incorrect
dirty state. This was introduced with commit c7adc610f0233b9afd5

```
    # pirate has_one ship
    ship = Ship.new(name: "Nights Dirty Lightning")
    pirate = ship.build_pirate(catchphrase: "Aye")
    ship.save!
    ship.previous_changes # => returns {} but this should contain the changes.
```

When saving ship the following happens:
1. the `before_save` callbacks of the ship are called
2. the callbacks call `autosave_associated_records_for_pirate`
3. `autosave_associated_records_for_pirate` saves the pirate
4. the `after_save` callbacks of the pirate are called
5. the callbacks call `autosave_associated_records_for_ship`
6. `autosave_associated_records_for_ship` saves the ship
7. the ship is saved again by the original save

`autosave_associated_records_for_ship` saves the ship because the
ship association is set by inverse_of in a `has_one` on the pirate. This does not happen
with a `has_many` because the inverse is not set.

With this commit the following happens:
1. the `before_save` callbacks of the ship are called
2. the callbacks call `autosave_associated_records_for_pirate`
3. set `@_already_called` on pirate for
  `autosave_associated_records_for_ship` to true
4. `autosave_associated_records_for_pirate` saves the pirate
5. the `after_save` callbacks of the pirate are called
6. the callbacks skip `autosave_associated_records_for_ship`
7. the ship is saved.