[documentation][ci skip] remove 'very receiver' from docs
This changes usage of 'very receiver' in the Active Support core extensions doc to just 'receiver'.  Throughout the rest of the Rails documentation, it is just 'receiver'.  'very' here is used (I think) to mean 'exact' in the sense of https://dictionary.cambridge.org/us/dictionary/english/very/#cacd-1-2-1 but it doesn't seem to add anything by its presence and may be confusing to some readers.