Schema Dumper raises an exception when it fails
Previoulsy the schema dumper when it encountered a failure it would
rescue the exception and leave a comment on the schema.rb. Now we raise
an exception and let the developer be aware that the schema was not
dumped sucessfuly. If an error is raised the `db/schema.rb` file remains
the same, we only modify it if the dump was successful.

Fixes: #36359