Add some examples for getting current_user
Add some examples for getting current_user in ActionCable::Connection.

It is somehow common for Rails to use a session in order to keep current user user_id. The most common authentication approach is with Devise.

This PR suggest some more info on how to obtain current_user with Devise and how to obtain current_user in a more general situation when we have a session cookie.

Also in the text a "signed" cookie is mentioned but in the code above is used an "encrypted" cookie.