Make getting/setting foreign_key uniform in autosave_association
Setting the foreign_key in the autosave records can be made more
similar.

Getting the foreign_key is done in two variants in autosave_association. The longer form:
```ruby
key = reflection.options[:primary_key] ? send(reflection.options[:primary_key]) : id
```

...and the shorter form:
```ruby
  key = send(reflection.options[:primary_key] || :id)
```

This refactoring uses the shorter form for both.

Also in save_belongs_to_association `association_id` is used as the name of the
variable for the value of the foreign_key. But `key` is used everywhere else in
autosave_association. So `association_id` has been renamed to `key`.