Add ActiveRecord::Base.current_env method
The current way for determining the environment is by doing something
`ActiveRecord::ConnectionHandling::DEFAULT_ENV.call` this method is
unclear what it's doing, and is a handful to type.

This patch introduces `ActiveRecord::Base.current_environment` this
method better expresses the intent and allows us to remove the
`current_environment` method from `MigrationContext` because we can call
this direcly.

Instead of having a couple of different places to get the current
environment from (connection_handler, migration_context) we can now
directly call this method internally within ActiveRecord.