Allow cache tests to be run in parallel
This initializes our caches with a random namespace which should ensure
that we can run the tests in parallel without interfering with each
other.

One compromise is that we can never actually call cache.clear on the
memcache client, since that would clear the cache for all namespaces. We
can work around this for the one test that needs to do a clear to
stub flush_all on the client and replace it with a delete.