Remove unused `DepthFirst` visitor
We only use `ToSQL` visitors in the our codebase, do not use
`DepthFirst` and `Dot` visitors.

The `DepthFirst` visitor (which was introduced at c86c37e5f) is used to
traverse an Arel (partial) ast with depth first.

Is there any worth to keep that undocumented feature with much code and
test cases.

This removes that unused `DepthFirst` code and test cases.