Lightly tweak Rails welcome index view
Yay, you're on Rails – When you boot up a blank Rails app for the first time, you see that message. It makes me so happy every time I see it. That said, the styling of that page perhaps leaves a little to be desired.

At the moment, we use Georgia as a font, with an *amazing* illustration, and some debug info underneath. This is fine.

This aims to:
- Modernize the CSS of the holding page to use flexbox, allowing the content to float in the centre on appropriate platforms
- Simultaneously, avoid compromising the experience in browsers without flexbox support
- Adopt system fonts for all typography rather than Georgia, et al
- Make things generally a little more aesthetically pleasing